Renaming 'interfacing' module.
