+++ tripleo/insights/git.py
