+++ tripleo/insights/lookup.py
@@ -19,7 +19,7 @@ import yaml
 
 from tripleo.insights.deployment import FeatureSetInterpreter
 from tripleo.insights.exceptions import InvalidURL
-from tripleo.insights.interfacing import GitDownloader, get_downloaders_for
+from tripleo.insights.git import GitDownloader, get_downloaders_for
 from tripleo.insights.io import DeploymentOutline, DeploymentSummary
 from tripleo.insights.validation import OutlineValidator
 from tripleo.utils.types import URL
